+++ RELEASE_NOTES.md
@@ -1,7 +1,9 @@
 # Release notes
 
-## New in git-machete 3.36.5
+## New in git-machete 3.37.0
 
+- added: `--start-from` flag of `traverse` now accepts branch names in addition to special values `here`, `root`, `first-root` (suggested by @chriscz)
+- improved: `--start-from` and `--return-to` flags of `traverse` also accept uppercase values like `HERE` or `STAY`
 - fixed: `fish` completion for long options with an argument
 
 ## New in git-machete 3.36.4